requirements.txt
pandas==1.4.2
requests==2.27.1
structlog==21.5.0
flask==2.1.1
psycopg2-binary==2.9.3
waitress==2.1.1
